+++ offchain/crates/solana-sdk/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod passport;
+pub mod reservation;
 pub mod revenue_distribution;
 
 //